+++ tests/integration/test_image.py
@@ -3,20 +3,114 @@
 
 """Image test module."""
 
+import glob
 import logging
+
+# Subprocess is used to run the application.
+import subprocess  # nosec: B404
 from pathlib import Path
 
 import pytest
+import pytest_asyncio
 from click.testing import CliRunner
 from fabric.connection import Connection as SSHConnection
+from openstack.compute.v2.image import Image
+from openstack.compute.v2.server import Server
 from openstack.connection import Connection
+from openstack.network.v2.security_group import SecurityGroup
 from pylxd import Client
 
 from github_runner_image_builder.cli import get_latest_build_id
 from github_runner_image_builder.config import IMAGE_OUTPUT_PATH
-from tests.integration import commands, helpers
+from tests.integration import commands, helpers, types
 
 logger = logging.getLogger(__name__)
+
+
+@pytest_asyncio.fixture(scope="module", name="openstack_server")
+async def openstack_server_fixture(
+    openstack_metadata: types.OpenstackMeta,
+    openstack_security_group: SecurityGroup,
+    openstack_image_name: str,
+    test_id: str,
+):
+    """A testing openstack instance."""
+    server_name = f"test-server-{test_id}"
+    images: list[Image] = openstack_metadata.connection.search_images(openstack_image_name)
+    assert images, "No built image found."
+    for server in helpers.create_openstack_server(
+        openstack_metadata=openstack_metadata,
+        server_name=server_name,
+        image=images[0],
+        security_group=openstack_security_group,
+    ):
+        yield server
+
+    for image in images:
+        openstack_metadata.connection.delete_image(image.id)
+
+
+@pytest_asyncio.fixture(scope="module", name="ssh_connection")
+async def ssh_connection_fixture(
+    openstack_server: Server,
+    proxy: types.ProxyConfig,
+    openstack_metadata: types.OpenstackMeta,
+    dockerhub_mirror: str | None,
+) -> SSHConnection:
+    """The openstack server ssh connection fixture."""
+    logger.info("Setting up SSH connection.")
+    ssh_connection = await helpers.wait_for_valid_connection(
+        connection=openstack_metadata.connection,
+        server_name=openstack_server.name,
+        network=openstack_metadata.network,
+        ssh_key=openstack_metadata.ssh_key.private_key,
+        proxy=proxy,
+        dockerhub_mirror=dockerhub_mirror,
+    )
+
+    return ssh_connection
+
+
+@pytest.fixture(scope="module", name="cli_run")
+def cli_run_fixture(
+    image: str,
+    cloud_name: str,
+    callback_script: Path,
+    openstack_connection: Connection,
+    openstack_image_name: str,
+):
+    """A CLI run.
+
+    This fixture assumes pipx is installed in the system and the github-runner-image-builder has
+    been installed using pipx. See testenv:integration section of tox.ini.
+    """
+    # This is a locally built application - we can trust it.
+    subprocess.check_call(  # nosec: B603
+        ["/usr/bin/sudo", Path.home() / ".local/bin/github-runner-image-builder", "init"]
+    )
+    subprocess.check_call(  # nosec: B603
+        [
+            "/usr/bin/sudo",
+            Path.home() / ".local/bin/github-runner-image-builder",
+            "run",
+            cloud_name,
+            openstack_image_name,
+            "--base-image",
+            image,
+            "--keep-revisions",
+            "2",
+            "--callback-script",
+            str(callback_script.absolute()),
+        ]
+    )
+
+    yield
+
+    openstack_image: Image
+    for openstack_image in openstack_connection.search_images(openstack_image_name):
+        openstack_connection.delete_image(openstack_image.id)
+    for image_file in glob.glob("*.img"):
+        Path(image_file).unlink(missing_ok=True)
 
 
 @pytest.mark.asyncio
