+++ tests/integration/test_openstack_builder.py
@@ -12,13 +12,13 @@ import logging
 import typing
 from datetime import datetime, timezone
 
-import openstack
 import pytest
 import pytest_asyncio
 from fabric.connection import Connection as SSHConnection
 from openstack.compute.v2.image import Image
 from openstack.compute.v2.server import Server
 from openstack.connection import Connection
+from openstack.network.v2.security_group import SecurityGroup
 
 from github_runner_image_builder import config, openstack_builder
 from tests.integration import helpers, types
@@ -71,31 +71,12 @@ def test_initialize(openstack_connection: Connection, arch: config.Arch, cloud_n
     assert openstack_connection.get_keypair(name_or_id=openstack_builder.BUILDER_SSH_KEY_NAME)
 
 
-class OpenStackMetadata(typing.NamedTuple):
-    """The OpenStack metadata wrapper fixture.
-
-    Attributes:
-        cloud_name: The OpenStack cloud name to use.
-        network: The OpenStack network to use.
-        flavor: The OpenStack flavor to use.
-    """
-
-    cloud_name: str
-    network: str
-    flavor: str
-
-
-@pytest.fixture(scope="module", name="openstack_metadata")
-def openstack_metadata_fixture(cloud_name: str, network_name: str, flavor_name: str):
-    """A wrapper fixture containing OpenStack metadata."""
-    return OpenStackMetadata(cloud_name=cloud_name, network=network_name, flavor=flavor_name)
-
-
 @pytest.fixture(scope="module", name="cli_run")
 def cli_run_fixture(
     arch: config.Arch,
     image: str,
-    openstack_metadata: OpenStackMetadata,
+    cloud_name: str,
+    openstack_metadata: types.OpenstackMeta,
     proxy: types.ProxyConfig,
 ):
     """A CLI run.
@@ -107,7 +88,7 @@ def cli_run_fixture(
         arch=arch,
         base=config.BaseImage.from_str(image),
         cloud_config=openstack_builder.CloudConfig(
-            cloud_name=openstack_metadata.cloud_name,
+            cloud_name=cloud_name,
             flavor=openstack_metadata.flavor,
             network=openstack_metadata.network,
         ),
@@ -116,46 +97,32 @@ def cli_run_fixture(
     )
 
 
-# The code is not duplicated, it has similar setup but uses different input fixtures for external
-# openstack builder.
+# the code is similar but the fixture source is localized and is different.
 # pylint: disable=R0801
-@pytest.mark.usefixtures("cli_run")
-@pytest.fixture(scope="module", name="server")
-def server_fixture(
+@pytest_asyncio.fixture(scope="module", name="openstack_server")
+async def openstack_server_fixture(
     openstack_metadata: types.OpenstackMeta,
+    openstack_security_group: SecurityGroup,
     test_id: str,
 ):
-    """The OpenStåck sserver fixture."""
+    """A testing openstack instance."""
     image: Image = openstack_metadata.connection.get_image(
         name_or_id=openstack_builder.IMAGE_SNAPSHOT_NAME
     )
     server_name = f"test-image-builder-run-{test_id}"
-    try:
-        server: Server = openstack_metadata.connection.create_server(
-            name=server_name,
-            image=image.id,
-            key_name=openstack_builder.BUILDER_SSH_KEY_NAME,
-            flavor=openstack_metadata.flavor,
-            network=openstack_metadata.network,
-            security_groups=[openstack_builder.SHARED_SECURITY_GROUP_NAME],
-            auto_ip=False,
-            timeout=5 * 60,
-            wait=True,
-        )
+    for server in helpers.create_openstack_server(
+        openstack_metadata=openstack_metadata,
+        server_name=server_name,
+        image=image,
+        security_group=openstack_security_group,
+    ):
         yield server
-    except openstack.exceptions.SDKException:
-        server = openstack_metadata.connection.get_server(name_or_id=server_name)
-        logger.exception("Failed to create server, %s", dict(server))
-    finally:
-        openstack_metadata.connection.delete_server(server_name, wait=True)
-        openstack_metadata.connection.delete_image(
-            openstack_builder.IMAGE_SNAPSHOT_NAME, wait=True
-        )
+    openstack_metadata.connection.delete_image(image.id)
 
 
 @pytest_asyncio.fixture(scope="module", name="ssh_connection")
 async def ssh_connection_fixture(
-    server: Server,
+    openstack_server: Server,
     proxy: types.ProxyConfig,
     openstack_metadata: types.OpenstackMeta,
     dockerhub_mirror: str | None,
@@ -164,7 +131,7 @@ async def ssh_connection_fixture(
     logger.info("Setting up SSH connection.")
     ssh_connection = await helpers.wait_for_valid_connection(
         connection=openstack_metadata.connection,
-        server_name=server.name,
+        server_name=openstack_server.name,
         network=openstack_metadata.network,
         ssh_key=openstack_metadata.ssh_key.private_key,
         proxy=proxy,
